Maya/qc status chips (#1554)
* fewer background api calls make the app run smoothly and reduce unnecessary stress on the BE

* minimum size for cells is 50 instead of 150

* spacing for tree icon

* introduce enterdelay for table header tooltips

* shift header select all checkbox to align with text

* fix lineage cells and tooltips

* truncate headers

* fix upload date today filter

* types

* types

* pull out qc chip into own component

* use new qc chip in new samples table

* types
