+++ src/frontend/src/views/Data/cellRenderers.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/display-name */
 
-import { ChipProps, Icon } from "czifui";
+import { Icon } from "czifui";
 import {
   defaultSampleCellRenderer,
   defaultTreeCellRenderer,
@@ -10,12 +10,11 @@ import {
   RowContent,
   TreeRowContent,
 } from "src/common/components/library/data_table/style";
-import { NewTabLink } from "src/common/components/library/NewTabLink";
 import { createTableCellRenderer } from "src/common/utils";
 import { datetimeWithTzToLocalDate } from "src/common/utils/timeUtils";
 import { CZ_BIOHUB_GROUP } from "src/views/Data/constants";
 import { LineageTooltip } from "./components/SamplesView/components/SamplesTable/components/LineageTooltip";
-import { StatusChip } from "./components/StatusChip";
+import { QualityScoreTag } from "./components/SamplesView/components/SamplesTable/components/QualityScoreTag";
 import { TreeActionMenu } from "./components/TreesView/components/TreesTable/components/TreeActionMenu";
 import TreeTableNameCell from "./components/TreesView/components/TreesTable/components/TreeTableNameCell";
 import { TreeTypeTooltip } from "./components/TreesView/components/TreesTable/components/TreeTypeTooltip";
@@ -30,32 +29,6 @@ import {
   UnderlinedRowContent,
 } from "./style";
 
-const LABEL_STATUS: Record<
-  string,
-  { label: string; status: NonNullable<ChipProps["status"]> }
-> = {
-  error: {
-    label: "bad",
-    status: "error",
-  },
-  success: {
-    label: "good",
-    status: "success",
-  },
-  warning: {
-    label: "mediocre",
-    status: "warning",
-  },
-  processing: {
-    label: "processing",
-    status: "pending",
-  },
-  failed: {
-    label: "failed",
-    status: "pending",
-  },
-};
-
 const PrivateId = ({
   value,
   item,
@@ -64,69 +37,6 @@ const PrivateId = ({
   item: Sample;
 }): JSX.Element => {
   const { qcMetrics, private: isPrivate, submittingGroup, uploadedBy } = item;
-
-  const PROCESSING_STATUS_TOOLTIP_TEXT = (
-    <div>
-      <div>
-        This sample doesn’t currently have a quality score because it is still
-        processing. Score will update when complete.
-      </div>
-    </div>
-  );
-
-  const GENERIC_STATUS_TOOLTIP_TEXT = (
-    <div>
-      <div>
-        <b>Quality Score: </b> Overall QC score from Nextclade which considers
-        genome completion and screens for potential contamination and sequencing
-        or bioinformatics errors.{" "}
-        <NewTabLink
-          href={
-            "https://docs.nextstrain.org/projects/nextclade/en/stable/user/algorithm/07-quality-control.html"
-          }
-        >
-          Learn more
-        </NewTabLink>
-      </div>
-    </div>
-  );
-
-  const FAILED_STATUS_TOOLTIP_TEXT = (
-    <div>
-      <div>
-        QC may fail when the sequence processing failed due to quality, or if
-        the uploaded sequence does not represent the correct pathogen
-      </div>
-    </div>
-  );
-
-  const labelQCStatus = () => {
-    const qcStatus = qcMetrics[0]?.qc_status;
-    switch (qcStatus.toLowerCase()) {
-      case "good":
-        return LABEL_STATUS.success;
-      case "bad":
-        return LABEL_STATUS.error;
-      case "mediocre":
-        return LABEL_STATUS.warning;
-      case "failed":
-        return LABEL_STATUS.failed;
-      default:
-        return LABEL_STATUS.processing;
-    }
-  };
-
-  const qcStatusLabel = labelQCStatus();
-
-  const LABEL_TO_TOOLTIP_TEXT: Record<string, JSX.Element> = {
-    processing: PROCESSING_STATUS_TOOLTIP_TEXT,
-    bad: GENERIC_STATUS_TOOLTIP_TEXT,
-    good: GENERIC_STATUS_TOOLTIP_TEXT,
-    mediocre: GENERIC_STATUS_TOOLTIP_TEXT,
-    failed: FAILED_STATUS_TOOLTIP_TEXT,
-  };
-
-  const { label, status } = qcStatusLabel;
 
   const displayName =
     submittingGroup?.name === CZ_BIOHUB_GROUP ? "CZ Biohub" : uploadedBy?.name;
@@ -154,11 +64,7 @@ const PrivateId = ({
         <PrivateIdValueWrapper>
           <CenteredFlexContainer>
             <span data-test-id="row-private-id">{value}</span>
-            <StatusChip
-              label={label}
-              status={status}
-              tooltipText={LABEL_TO_TOOLTIP_TEXT[label]}
-            />
+            <QualityScoreTag qcMetric={qcMetrics[0]} />
           </CenteredFlexContainer>
           <StyledUploaderName data-test-id="row-user-name">
             {displayName}
